# swagger.json
{
    "openapi": "3.0.1",
    "info": {
      "version": "1.0.0",
      "title": "Plainx middleware api documentation",
      "description": "API documentation for plainx middleware",
      "license": {
        "name": "MIT",
        "url": "https://opensource.org/licenses/MIT"
      }
    },
    "host": "https://plainx-midddleware-engine.azurewebsites.net",
    "basePath": "/",
    "components": {
        "securitySchemes": {
          "jwt": {
            "type": "http",
            "scheme": "bearer",
            "name": "x-access-token",
            "in": "header",
            "bearerFormat": "JWT"
          }
        }
      },
      "security": [{
        "jwt": []
      }],
...
    "schemes": [
      "https"
    ],
    "consumes": [
      "application/json"
    ],
    "produces": [
      "application/json"
    ],
    "paths": {
      "/user": {
        "post": {
          "tags": [
            "Users"
          ],
          "parameters": [
            {
              "name": "user",
              "in": "body",
              "description": "User that we want to create",
              "schema": {
                "$ref": "#/definitions/User"
              }
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/User"
                }
              }
            }
          },
          "produces": [
            "application/json"
          ],
          "responses": {
            "200": {
              "description": "New user is created",
              "schema": {
                "$ref": "#/definitions/User"
              }
            },
            "401": {
                "description": "Email address must be unique",
                "schema": {
                    "$ref": "#/definitions/User"
                }
            }
          }
        },
        "get": {
          "tags": [
            "Users"
          ],
          "summary": "Get all users in system",
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/Users"
              }
            },
            "403": {
              "description": "Failed to authenticate token."
            }
          }
        }
      },
...
      "/user/forgot_password":{
        "post": {
          "tags": [
            "Users"
          ],
          "summary": "Send Password reset link",
          "parameters": [
            {
              "name": "Password",
              "in": "body",
              "description": "Login Id that require password change",
              "schema": {
                  "required": [
                    "email",
                    "resetPasswordEndPoint"
                  ],
                  "properties": {
                      "email": { "type": "string", "uniqueItems": true },
                      "resetPasswordEndPoint":{ "type": "string" }
                  }
              }
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "required": [
                    "email",
                    "resetPasswordEndPoint"
                  ],
                  "properties": {
                      "email": { "type": "string", "uniqueItems": true },
                      "resetPasswordEndPoint":{ "type": "string" }
                  }
                }
              }
            }
          },
          "produces": [
            "application/json"
          ],
          "responses": {
            "200": {
              "description": "Password reset link has been sent to your email account",
              "schema": {
                "required": [
                  "email",
                  "resetPasswordEndPoint"
                ],
                "properties": {
                    "email": { "type": "string", "uniqueItems": true },
                    "resetPasswordEndPoint":{ "type": "string" }
                }
              }
            },
            "500": {
                "description": "An error occured with the system, Please try after sometime",
                "schema": {
                  "required": [
                    "email",
                    "resetPasswordEndPoint"
                  ],
                  "properties": {
                      "email": { "type": "string", "uniqueItems": true },
                      "resetPasswordEndPoint":{ "type": "string" }
                  }
                }
            },
            "400": {
                "description": "User with given email doesn't exist",
                "schema": {
                  "required": [
                    "email",
                    "resetPasswordEndPoint"
                  ],
                  "properties": {
                      "email": { "type": "string", "uniqueItems": true },
                      "resetPasswordEndPoint":{ "type": "string" }
                  }
                }
            }
          }
        }
      },
      "/user/update_password/{userId}/{token}":{
        "post": {
          "tags": [
            "Users"
          ],
          "summary": "Updates Password",
          "parameters": [
            {
              "name": "userId",
              "in": "path",
              "required": true,
              "description": "ID of user that we want to Update Password",
              "type": "string"
            },
            {
              "name": "token",
              "in": "path",
              "required": true,
              "description": "token",
              "type": "string"
            },
            {
              "name": "password",
              "in": "body",
              "required": true,
              "description": "password",
              "schema": {
                "required": [
                  "password"
                ],
                "properties": {
                    "password": { "type": "string", "uniqueItems": true }
                }
              }
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "required": [
                    "password"
                  ],
                  "properties": {
                      "password": { "type": "string", "uniqueItems": true }
                  }
                }
              }
            }
          },
          "produces": [
            "application/json"
          ],
          "responses": {
            "200": {
              "description": "password reset sucessfully.",
              "schema": {
                "required": [
                  "password"
                ],
                "properties": {
                    "password": { "type": "string", "uniqueItems": true }
                }
              }
            },
            "500": {
                "description": "An error occured with the system, Please try after sometime",
                "schema": {
                  "required": [
                    "password"
                  ],
                  "properties": {
                      "password": { "type": "string", "uniqueItems": true }
                  }
                }
            },
            "400": {
                "description": "Please enter valid password OR Invalid link or expired",
                "schema": {
                  "required": [
                    "password"
                  ],
                  "properties": {
                      "password": { "type": "string", "uniqueItems": true }
                  }
                }
            }
          }
        }
      },
      "/user/verify_token":{
        "post": {
          "tags": [
            "Users"
          ],
          "summary": "Verify the JWT Token",
          "parameters": [
            {
              "name": "Verify Token",
              "in": "body",
              "description": "Verify the JWT Token",
              "schema": {
                "required": [
                  "userId"
                ],
                "properties": {
                    "userId": { "type": "string", "uniqueItems": true }
                }
              }
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "required": [
                    "userId"
                  ],
                  "properties": {
                      "userId": { "type": "string", "uniqueItems": true }
                  }
                }
              }
            }
          },
          "produces": [
            "application/json"
          ],
          "responses": {
            "200": {
              "description": "Token Verified",
              "schema": {
                "required": [
                  "userId"
                ],
                "properties": {
                    "userId": { "type": "string", "uniqueItems": true }
                }
              }
            },
            "500": {
                "description": "There was a problem finding the user.",
                "schema": {
                  "required": [
                    "userId"
                  ],
                  "properties": {
                      "userId": { "type": "string", "uniqueItems": true }
                  }
                }
            },
            "404": {
                "description": "No user found.",
                "schema": {
                  "required": [
                    "userId"
                  ],
                  "properties": {
                      "userId": { "type": "string", "uniqueItems": true }
                  }
                }
            }
          }
        }
      },
      "/user/{userId}": {
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "description": "ID of user that we want to find",
            "type": "string"
          }
        ],
        "get": {
          "tags": [
            "Users"
          ],
          "summary": "Get user with given ID",
          "responses": {
            "200": {
              "description": "User is found",
              "schema": {
                "$ref": "#/definitions/User"
              }
            },
            "403": {
              "description": "Failed to authenticate token."
            }
          }
        },
        "delete": {
          "summary": "Delete user with given ID",
          "tags": [
            "Users"
          ],
          "responses": {
            "200": {
              "description": "User is deleted",
              "schema": {
                "$ref": "#/definitions/User"
              }
            }
          }
        },
        "put": {
          "summary": "Update user with give ID",
          "tags": [
            "Users"
          ],
          "parameters": [
            {
              "name": "user",
              "in": "body",
              "description": "User with new values of properties",
              "schema": {
                "$ref": "#/definitions/User"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "User is updated",
              "schema": {
                "$ref": "#/definitions/User"
              }
            }
          }
        }
      },
...
    },
    "definitions": {
      "User": {
        "required": [
          "email"
        ],
        "properties": {
            "firstName": { "type": "string" },
            "lastName": { "type": "string"},
            "email": { "type": "string", "uniqueItems": true },
            "password": { "type": "string"},
            "company": { "type": "string"},
            "designation": { "type": "string" }
        }
      },
      "Users": {
        "type": "array",
        "$ref": "#/definitions/User"
      },
...
    }
  }